+++ CHANGELOG.md
@@ -6,8 +6,8 @@
 - **REFACTOR**: Update default `anchor` of `ShadSelect` from `ShadAnchorAuto()` to `ShadAnchorAuto(offset: Offset(0, 4))`.
 - **REFACTOR**: Add `checkboxPadding` to `ShadCheckbox`, `ShadCheckboxTheme` and `ShadCheckboxFormField`, defaults to `EdgeInsets.only(top: 1)`.
 - **REFACTOR**: Add `radioPadding` to `ShadRadio` and `ShadRadioTheme`, defaults to `EdgeInsets.only(top: 1)`.
-- **REFACTOR**: The default offset of `ShadAnchorAuto` has been changed from `Offset.zero` to `Offset(0, 4)`.
 - **FIX**: Background color of pages for pure `ShadApp` if no `Scaffold is used`, by adding a new `backgroundColor` parameter to `ShadApp` and `ShadAppBuilder`.
+- **REFACTOR**: Update the `ShadTooltip` default anchor, defaults to `Offset(0, -4)` instead of `Offset.zero`.
 
 ## 0.26.5